Create compare view utils.js file
- create compare view utils.js file
- move hydrology scenario mapping code into utils
